requirements.txt
requests
groq
pytz
google-generativeai
Flask